Make websocket logic extractable
