+++ packages/stream_video/lib/src/coordinator/open_api/coordinator_ws.dart
@@ -5,13 +5,13 @@ import 'package:internet_connection_checker_plus/internet_connection_checker_plu
 
 import '../../../globals.dart';
 import '../../../open_api/video/coordinator/api.dart' as open;
+import '../../core/network_monitor_flutter.dart';
 import '../../core/video_error.dart';
 import '../../logger/impl/tagged_logger.dart';
 import '../../models/user_info.dart';
 import '../../retry/retry_policy.dart';
 import '../../shared_emitter.dart';
 import '../../token/token_manager.dart';
-import '../../types/other.dart';
 import '../../utils/none.dart';
 import '../../utils/result.dart';
 import '../../ws/health/health_monitor.dart';
@@ -34,9 +34,7 @@ String _buildUrl(String baseUrl, String apiKey) {
       '&X-Stream-Client=$xStreamClientHeader';
 }
 
-class CoordinatorWebSocket extends StreamWebSocket
-    with ConnectionStateMixin
-    implements HealthListener {
+class CoordinatorWebSocket extends StreamWebSocket implements HealthListener {
   CoordinatorWebSocket(
     String url, {
     Iterable<String>? protocols,
@@ -57,7 +55,8 @@ class CoordinatorWebSocket extends StreamWebSocket
   late final HealthMonitor healthMonitor = HealthMonitorImpl(
     'Coord',
     this,
-    networkMonitor: networkMonitor,
+    networkMonitor:
+        NetworkMonitorFlutter.fromInternetConnection(networkMonitor),
   );
 
   /// The API key used to authenticate the user.
