+++ packages/stream_video/lib/src/ws/health/network_monitor.dart
@@ -1,0 +1,11 @@
+abstract class NetworkMonitor {
+  Stream<NetworkStatus> get onStatusChange;
+}
+
+enum NetworkStatus {
+  /// Internet is available because at least one of the HEAD requests succeeded.
+  connected,
+
+  /// None of the HEAD requests succeeded. Basically, no internet.
+  disconnected,
+}
